+++ Assets/Scripts/src/NetStack.Serialization/SevenBitDecoding.cs
@@ -59,7 +59,7 @@ namespace NetStack.Serialization
         public i32 i32(RawBitReader<TMemory> b) => decode(u32(b));
 
         [MethodImpl(Optimization.AggressiveInliningAndOptimization)]
-        public i32 i32(RawBitReader<TMemory> b, i32 numberOfBits)
+        public i32 i32(RawBitReader<TMemory> b, u8 numberOfBits)
         {
             u32 value = b.u32(numberOfBits);
             return decode(value);
